Navigation, font and moved some components
