+++ src/app/reducers.js
@@ -1,10 +1,6 @@
 import { combineReducers } from 'redux';
-import counter from '../counter/reducer';
-import beer from '../beer/reducer';
-import stopList from '../stop-list/reducer';
+import stopList from '../components/stop-list/reducer';
 
 export default combineReducers({
-  counter,
-  beer,
   stopList,
 });
