+++ src/app/root.jsx
@@ -2,17 +2,13 @@ import Inferno from 'inferno';
 import { Provider } from 'inferno-redux';
 import { Router, Route, IndexRoute } from 'inferno-router';
 import App from './app.jsx';
-import Beer from '../beer/beer.jsx';
-import Counter from '../counter/counter.jsx';
-import StopList from '../stop-list/stop-list.jsx';
+import StopList from '../components/stop-list/stop-list.jsx';
 
 const Root = ({ store, history }) => (
   <Provider store={store}>
     <Router history={ history }>
       <Route component={ App }>
-        <IndexRoute component={ Beer }/>
-        <Route path="counter" component={ Counter } />
-        <Route path="stops" component={ StopList } />
+        <IndexRoute component={ StopList }/>
       </Route>
     </Router>
   </Provider>
