+++ src/app/sagas.js
@@ -1,10 +1,8 @@
 import { fork } from 'redux-saga/effects';
-import beer from '../beer/saga';
-import stopList from '../stop-list/saga';
+import stopList from '../components/stop-list/saga';
 
 export default function* rootSaga() {
   yield [
-    fork(beer),
     fork(stopList),
   ];
 }
