+++ src/app/styles.js
@@ -1,5 +1,4 @@
 // Import styling to this file
-import '../beer/style.scss';
-import '../counter/style.scss';
-import '../stop-list/style.scss';
-import '../stop/style.scss';
+import '../components/stop-list/style.scss';
+import '../components/stop/style.scss';
+import '../components/navigation/style.scss';
