--- src/beer/actions.js
@@ -1,3 +1,0 @@
-export const FETCH_BEER = 'FETCH_BEER';
-export const FETCH_BEER_SUCCESS = 'FETCH_BEER_SUCCESS';
-export const FETCH_BEER_FAILED = 'FETCH_BEER_FAILED';
