--- src/beer/style.scss
@@ -1,3 +1,0 @@
-.beer {
-  
-}
