+++ src/components/stop-list/reducer.js
