+++ src/components/stop-list/style.scss
@@ -2,4 +2,5 @@
   margin: 20px;
   display: flex;
   flex-wrap: wrap;
+  justify-content: center;
 }
