+++ src/components/stop/actions.js
