+++ src/components/stop/reducer.js
