+++ src/components/stop/style.scss
