--- src/counter/actions.js
@@ -1,2 +1,0 @@
-export const INCREASE = 'INCREASE';
-export const DECREASE = 'DECREASE';
