--- src/counter/style.scss
@@ -1,3 +1,0 @@
-.counter {
-  background-color: gray;
-}
